refactor: restructure MarketDetailV2 header and component imports
- Introduced a new MarketDetailHeader component to encapsulate header logic.
- Simplified the MarketDetail component by removing unused imports and JSX elements.
- Enhanced code organization by updating component exports in the index file.
