+++ packages/kit/src/views/Market/MarketDetailV2/MarketDetailV2.tsx
@@ -1,33 +1,24 @@
-import { useEffect, useMemo } from 'react';
+import { useEffect } from 'react';
 
 import type { IPageScreenProps } from '@onekeyhq/components';
-import { NavBackButton, Page, XStack, useMedia } from '@onekeyhq/components';
+import { Page, useMedia } from '@onekeyhq/components';
 import { EJotaiContextStoreNames } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import {
   EAppEventBusNames,
   appEventBus,
 } from '@onekeyhq/shared/src/eventBus/appEventBus';
-import platformEnv from '@onekeyhq/shared/src/platformEnv';
-import { ETabRoutes } from '@onekeyhq/shared/src/routes';
 import type {
   ETabMarketRoutes,
   ITabMarketParamList,
 } from '@onekeyhq/shared/src/routes';
 import { EAccountSelectorSceneName } from '@onekeyhq/shared/types';
 
-import {
-  AccountSelectorProviderMirror,
-  AccountSelectorTriggerHome,
-} from '../../../components/AccountSelector';
-import { TabPageHeader } from '../../../components/TabPageHeader';
-import { TabPageHeaderContainer } from '../../../components/TabPageHeader/components/TabPageHeaderContainer';
-import { HeaderLeftCloseButton } from '../../../components/TabPageHeader/HeaderLeft';
+import { AccountSelectorProviderMirror } from '../../../components/AccountSelector';
 import { useTokenDetailActions } from '../../../states/jotai/contexts/marketV2';
 import { MarketWatchListProviderMirrorV2 } from '../MarketWatchListProviderMirrorV2';
 
-import { TokenDetailHeader } from './components/TokenDetailHeader/TokenDetailHeader';
+import { MarketDetailHeader } from './components/MarketDetailHeader';
 import { useAutoRefreshTokenDetail } from './hooks';
-import { useMarketDetailBackNavigation } from './hooks/useMarketDetailBackNavigation';
 import { DesktopLayout } from './layouts/DesktopLayout';
 import { MobileLayout } from './layouts/MobileLayout';
 
@@ -37,7 +28,6 @@ function MarketDetail({
   const { tokenAddress, networkId } = route.params;
   const media = useMedia();
   const tokenDetailActions = useTokenDetailActions();
-  const { handleBackPress } = useMarketDetailBackNavigation();
 
   // Clear all token detail content when unmount
   useEffect(() => {
@@ -56,47 +46,9 @@ function MarketDetail({
     networkId,
   });
 
-  const customHeaderLeft = useMemo(
-    () => (
-      <XStack gap="$3" ai="center">
-        <NavBackButton onPress={handleBackPress} />
-        <AccountSelectorTriggerHome num={0} />
-      </XStack>
-    ),
-    [handleBackPress],
-  );
-
-  const customHeaderRight = useMemo(() => null, []);
-
   return (
     <Page>
-      {platformEnv.isNative ? (
-        <TabPageHeaderContainer>
-          <XStack gap="$3" ai="center">
-            <HeaderLeftCloseButton />
-
-            <MarketWatchListProviderMirrorV2
-              storeName={EJotaiContextStoreNames.marketWatchListV2}
-            >
-              <TokenDetailHeader
-                containerProps={{ p: '$0' }}
-                showStats={false}
-                showMediaAndSecurity={false}
-              />
-            </MarketWatchListProviderMirrorV2>
-          </XStack>
-        </TabPageHeaderContainer>
-      ) : (
-        <TabPageHeader
-          sceneName={EAccountSelectorSceneName.home}
-          tabRoute={ETabRoutes.Market}
-          customHeaderLeftItems={customHeaderLeft}
-          customHeaderRightItems={
-            platformEnv.isNative ? customHeaderRight : null
-          }
-          hideSearch={!media.gtMd}
-        />
-      )}
+      <MarketDetailHeader />
 
       <Page.Body>{media.gtLg ? <DesktopLayout /> : <MobileLayout />}</Page.Body>
     </Page>
