+++ packages/kit/src/views/Market/MarketDetailV2/components/MarketDetailHeader/index.ts
@@ -1,0 +1,1 @@
+export { MarketDetailHeader } from './MarketDetailHeader';
