+++ packages/kit/src/views/Market/MarketDetailV2/components/index.ts
@@ -1,4 +1,5 @@
 export * from './InformationTabs';
+export * from './MarketDetailHeader';
 export * from './SwapPanel';
 export * from './TokenDetailHeader';
 export * from './TokenPriceChart';
